requirements.txt
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
cmdstanpy==1.2.5
colorama==0.4.6
comm==0.2.3
contourpy==1.3.3
curl_cffi==0.13.0
cycler==0.12.1
debugpy==1.8.16
decorator==5.2.1
docopt==0.6.2
executing==2.2.1
fonttools==4.59.2
frozendict==2.4.6
gitdb==4.0.12
GitPython==3.1.45
holidays==0.80
idna==3.10
importlib_resources==6.5.2
ipykernel==6.30.1
ipython==9.5.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.4.9
lxml==6.0.1
MarkupSafe==3.0.2
matplotlib==3.10.6
matplotlib-inline==0.1.7
multitasking==0.0.12
narwhals==2.3.0
nest-asyncio==1.6.0
numpy==2.3.2
packaging==25.0
pandas==2.3.2
parso==0.8.5
patsy==1.0.1
peewee==3.18.2
pillow==11.3.0
pipreqs==0.4.13
platformdirs==4.4.0
plotly==6.3.0
plotly-express==0.4.1
prompt_toolkit==3.0.52
protobuf==6.32.0
psutil==7.0.0
pure_eval==0.2.3
pyarrow==21.0.0
pycparser==2.22
pydeck==0.9.1
Pygments==2.19.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
pywin32==311
pyzmq==27.0.2
referencing==0.36.2
requests==2.32.5
rpds-py==0.27.1
scikit-learn==1.7.1
scipy==1.16.1
seaborn==0.13.2
six==1.17.0
smmap==5.0.2
soupsieve==2.8
stack-data==0.6.3
stanio==0.5.1
statsmodels==0.14.5
streamlit==1.50.0
tenacity==9.1.2
threadpoolctl==3.6.0
toml==0.10.2
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
watchdog==6.0.0
wcwidth==0.2.13
websockets==15.0.1
xgboost==3.0.5
yarg==0.1.10